Refactor: Change delete logic for batter security
Old delete logic had a vulnerability to IDOR attack.
New logic compare not only id of task, it also compare user_id from current session, if it equal task will be deleted.

+++ App/Models/Task.php
@@ -62,13 +62,16 @@ class Task
     }
 
     // Delete task by ID
-    public function delete($id)
+    public function delete($id, $userId)
     {
-        $query = "DELETE FROM tasks WHERE id = :id";
+        $query = "DELETE FROM tasks WHERE id = :id AND user_id = :user_id";
 
         $stmt = $this->db->prepare($query);
 
-        return $stmt->execute(['id' => $id]);
+        return $stmt->execute([
+            'id' => $id,
+            'user_id' => $userId
+            ]);
     }
 
     // Change is_completed status
